requirements.txt
discord.py
nekos.py
requests
random2